Refactor texture handling and block definitions
- Added new UV mapping arrays for handling texture rotations.
- Enhanced `addVoxelDataToChunk` to apply correct UV coordinates based on block orientation.
- Updated UV offset retrieval to use `textureId` from block's directional models.

+++ voxel game/BlockTexture.c
@@ -47,7 +47,7 @@ float GetUvOfOneBlockY()
 	return AtlasOneBlockInUvY;
 }
 
-void GetUvOffsetByTexId(int texId, float* x, float* y)
+void GetUvOffsetByTexId(uint16_t texId, float* x, float* y)
 {
 	int AtlasX = texId % AtlasWidthInBlocks;
 	int AtlasY = texId / AtlasWidthInBlocks;
